+++ src/controllers/product-controller.js
@@ -1,9 +1,5 @@
+import is from '@sindresorhus/is';
 const { productService } = require('../services/product-service');
-
-import {
-  ValueIsNullError,
-  CategoryDoesNotExistsError,
-} from '../error/value-error';
 
 class ProductController {
   async getProduct(req, res) {
@@ -23,11 +19,16 @@ class ProductController {
   }
 
   async addProduct(req, res, next) {
-    let imgUrl = null;
-    if (req.file) {
-      imgUrl = req.file.location;
-    }
     try {
+      if (is.emptyObject(req.body)) {
+        throw new Error(
+          'headers의 Content-Type을 application/json으로 설정해주세요',
+        );
+      }
+      let imgUrl = null;
+      if (req.file) {
+        imgUrl = req.file.location;
+      }
       const addedProduct = await productService.addProduct(req.body, imgUrl);
       return res.status(200).json({
         isSuccess: true,
@@ -41,11 +42,16 @@ class ProductController {
   }
 
   async editProduct(req, res, next) {
-    let imgUrl = null;
-    if (req.file) {
-      imgUrl = req.file.location;
-    }
     try {
+      if (is.emptyObject(req.body)) {
+        throw new Error(
+          'headers의 Content-Type을 application/json으로 설정해주세요',
+        );
+      }
+      let imgUrl = null;
+      if (req.file) {
+        imgUrl = req.file.location;
+      }
       const editedProduct = await productService.setProduct(req.body, imgUrl);
       return res.status(200).json({
         isSuccess: true,
@@ -58,8 +64,13 @@ class ProductController {
     }
   }
   async deleteProduct(req, res, next) {
-    const productId = req.body.productId;
     try {
+      if (is.emptyObject(req.body)) {
+        throw new Error(
+          'headers의 Content-Type을 application/json으로 설정해주세요',
+        );
+      }
+      const productId = req.body.productId;
       const deletedProduct = await productService.deleteProduct(productId);
       return res.status(200).json({
         isSuccess: true,
